[#2422] Refactor CLI commands: sync-billing-email and customers subcommand
- Rename sync_contact_email_command → sync_billing_email_command
- Update sync logic to use billing_email field instead of contact_email
- Add --update-contact-email option to optionally sync both fields
- Add --sleep option (default 50ms) to prevent Stripe API rate limiting
- Move sync_auth_accounts_command to customers/ subdirectory
- Update command registration to 'customers sync-auth-accounts'
- Fix SQLite-specific help text to be database-agnostic
