+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2025-12-06 06:37:56 UTC using RuboCop version 1.81.7.
+# on 2026-02-04 08:55:35 UTC using RuboCop version 1.81.7.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -38,13 +38,7 @@ Lint/RedundantSafeNavigation:
   Exclude:
     - 'apps/web/billing/cli/plans_validate_command.rb'
 
-# Offense count: 1
-# Configuration parameters: CountComments, Max, CountAsOne.
-Metrics/ClassLength:
-  Exclude:
-    - 'lib/onetime/redis_key_migrator.rb'
-
-# Offense count: 437
+# Offense count: 630
 # Configuration parameters: CountComments, CountAsOne, AllowedMethods, AllowedPatterns.
 Metrics/MethodLength:
   Max: 168
@@ -54,7 +48,6 @@ Metrics/MethodLength:
 Metrics/PerceivedComplexity:
   Exclude:
     - 'apps/api/domains/cli/info_command.rb'
-    - 'apps/web/auth/routes/account.rb'
     - 'apps/web/billing/cli/plans_validate_command.rb'
     - 'apps/web/billing/cli/prices_generate_command.rb'
     - 'apps/web/billing/cli/products_create_command.rb'
@@ -62,8 +55,8 @@ Metrics/PerceivedComplexity:
     - 'apps/web/billing/models/plan.rb'
     - 'apps/web/core/views/helpers/initialize_view_vars.rb'
     - 'lib/onetime/cli/change_email_command.rb'
+    - 'lib/onetime/cli/customers/sync_auth_accounts_command.rb'
     - 'lib/onetime/cli/migrate_redis_data_command.rb'
-    - 'lib/onetime/cli/sync_auth_accounts_command.rb'
     - 'lib/onetime/config.rb'
     - 'lib/onetime/initializers/print_log_banner.rb'
     - 'lib/onetime/utils.rb'
@@ -73,19 +66,12 @@ Security/MarshalLoad:
   Exclude:
     - 'lib/onetime/cli/session_command.rb'
 
-# Offense count: 3
+# Offense count: 1
 Style/MultilineBlockChain:
   Exclude:
-    - 'apps/api/colonel/logic/colonel/get_user_details.rb'
     - 'apps/web/billing/controllers/billing.rb'
 
-# Offense count: 1
-# Configuration parameters: Max.
-Style/SafeNavigationChainLength:
-  Exclude:
-    - 'lib/onetime/models/organization/features/with_organization_billing.rb'
-
-# Offense count: 112
+# Offense count: 125
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: AllowHeredoc, AllowURI, AllowQualifiedName, URISchemes, IgnoreCopDirectives, AllowedPatterns, SplitStrings.
 # URISchemes: http, https
