+++ lib/onetime/cli.rb
@@ -65,7 +65,7 @@ require_relative 'cli/server_command'
 require_relative 'cli/boot_test_command'
 require_relative 'cli/migrate_command'
 require_relative 'cli/migrate_redis_data_command'
-require_relative 'cli/sync_auth_accounts_command'
+require_relative 'cli/customers/sync_auth_accounts_command'
 require_relative 'cli/customers_command'
 require_relative 'cli/passwords_command'
 require_relative 'cli/test_data_command'
